Update to give reward by clicks and added random hold feature before force unlock
- TaskStateEnded: call give direct reward to deliver reward by clicks
instead of pause duration.
- GatingForceJudge: introduced a new field: timeRange, a uniformly
random value between (0,15), translates to 20-300ms in 20ms increments
only, a random hold period of force within the target range before
force unlock for free move. The target range is specified in the config
file, the same range previously used to determine if the force
threshold is reached.